# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Tnajjem API",
        "description": "Backend for ATGMO's Tnajjem application documented with Swagger",
        "license": {
            "name": "MIT",
            "url": "https://spdx.org/licenses/MIT.html"
        },
        "contact": {
            "name": "Raed Fadhlaoui",
            "url": "https://www.linkedin.com/in/fadhlaouiraed/",
            "email": "raed.fadhlaoui@orange.com"
        }
    },
    "host": "najjem.orange.tn",
    "basePath": "/v1/api",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/event/{Id}": {
            "parameters": [
                {
                    "name": "Id",
                    "in": "path",
                    "required": true,
                    "description": "ID of article that we want to find",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Event/News"
                ],
                "summary": "Get Event with given ID",
                "responses": {
                    "200": {
                        "description": "Event is found",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "403": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "summary": "Update Event with give ID",
                "tags": [
                    "Event/News"
                ],
                "parameters": [
                    {
                        "name": "Event",
                        "in": "body",
                        "description": "Event with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event is updated",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Event with given ID",
                "tags": [
                    "Event/News"
                ],
                "responses": {
                    "200": {
                        "description": "Event is deleted"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Event": {
            "properties": {
                "gallery": {
                    "type": "array",
                    "items": {}
                },
                "photo": {
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                },
                "end_at": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "isEvent": {
                    "type": "boolean"
                },
                "ar": {
                    "type": "object",
                    "properties": {
                        "event_name": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "event_location": {
                            "type": "string"
                        }
                    }
                },
                "fr": {
                    "type": "object",
                    "properties": {
                        "event_name": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "event_location": {
                            "type": "string"
                        }
                    }
                },
                "coordinates": {
                    "type": "object",
                    "properties": {
                        "latitude": {
                            "type": "string"
                        },
                        "longitude": {
                            "type": "string"
                        }
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}